Backend/remove-account/requirements.txt
functions-framework==3.*
google-cloud-firestore
firebase-admin